+++ app/definitions/pages/common/payee-details/about-the-person-or-company-being-paid.js
@@ -1,4 +1,5 @@
 /* eslint-disable no-console */
+const JourneyContext = require('@dwp/govuk-casa/lib/JourneyContext');
 const fieldValidators = require(
   '../../../field-validators/common/payee-details/about-the-person-or-company-being-paid',
 );
@@ -8,7 +9,6 @@ const checkIfShouldHaveOptionToAddNewPayeeThroughEditMode = (req, res) => {
   const { newPayee: wantsToAddNewPayee } = { ...req.casa.journeyContext.getDataForPage('__hidden_new_payee__') };
   const { account } = { ...req.casa.journeyContext.getDataForPage('__hidden_account__') };
   const hasExistingPayees = account.payees.length !== 0;
-
   res.locals.shouldHaveOptionToAddNewPayee = hasExistingPayees && !wantsToAddNewPayee;
 };
 
@@ -27,9 +27,32 @@ module.exports = () => ({
       if (res.locals.journeyType === claimTypesFullName.TW) {
         res.locals.howDidYouTravelForWork = req.casa.journeyContext.getDataForPage('which-journey-type').howDidYouTravelForWork;
       }
-      res.locals.payees = req.casa.journeyContext.getDataForPage('__hidden_account__').account.payees;
+
+      const payees = req.casa.journeyContext.getDataForPage('__hidden_account__').account.payees;
+      let filteredPayeeDetails;
+      filteredPayeeDetails = payees.filter((value, index, self) =>
+        index === self.findIndex((t) => (
+          t.name === value.name && t.emailAddress === value.emailAddress
+        ))
+      )
+      res.locals.payees = filteredPayeeDetails.slice(0, 10);
+
       checkIfShouldHaveOptionToAddNewPayeeThroughEditMode(req, res);
       next();
+    },
+    preredirect: (req, res, next) => {
+      if (req.inEditMode) {
+        JourneyContext.putContext(req.session, req.casa.journeyContext);
+
+        req.session.save((err) => {
+          if (err) {
+            throw err;
+          }
+          res.redirect(`select-person-company-being-paid?edit=&editorigin=${req.editOriginUrl}`);
+        });
+      } else {
+        next();
+      }
     },
     postvalidate: (req, res, next) => {
       const payeeValue = req.casa.journeyContext.getDataForPage('person-company-being-paid').payee;
@@ -49,10 +72,14 @@ module.exports = () => ({
           newPayee: false,
         });
 
-        req.casa.journeyContext.setDataForPage('__hidden_existing_payee__', {
+        // new hidden page called hidden_existing_payee_details so that data can be retained going forward
+        // even when pressing back button on existing bank details pages
+        req.casa.journeyContext.setDataForPage('__hidden_existing_payee_details__', {
+          id: payeeSelected.id,
           fullName: payeeSelected.name ?? payeeSelected.bankAccountName,
-          emailAddress: payeeSelected.email,
+          emailAddress: payeeSelected.emailAddress,
         });
+
         req.casa.journeyContext.setDataForPage('person-company-being-paid-details', undefined);
         req.casa.journeyContext.setValidationErrorsForPage('person-company-being-paid-details', undefined);
         req.casa.journeyContext.setDataForPage(
@@ -66,7 +93,8 @@ module.exports = () => ({
         req.casa.journeyContext.setDataForPage(
           'person-company-being-paid-address',
           undefined,
-        ); req.casa.journeyContext.setValidationErrorsForPage(
+        ); 
+        req.casa.journeyContext.setValidationErrorsForPage(
           'person-company-being-paid-address',
           undefined,
         );